Improve error messaging and add certificate expiration check
Refactored the error message when insufficient arguments are provided to clarify that a PEM filename is expected, and added a deferred error check for closing the file. Also included is checking and displaying certificate expiration details. A README.md file has also been created for a better overview and usage instructions of this program.